manfa/requirements.txt
torch>=2.1
mamba-ssm>=1.2.0
numpy>=1.25
scipy>=1.10
pyyaml>=6.0
networkx>=3.1
pandas>=2.0
tqdm>=4.66
